+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1
 
+* Add `mapMaybe` and `catMaybes`.
 * Add fusion rules for `genericTake`.
 * Remove harmful fusion rules for `drop` and `dropWhile`.
 * Fix `instance Monad Infinite` on 32-bit machines.